Force correct types for settings read from DB.
This should help eliminate the parade of issues we've had that stem
from settings not being converted to the proper type.

Fixes #382.

+++ src/local_deep_research/config/llm_config.py
@@ -79,7 +79,7 @@ def get_llm(model_name=None, temperature=None, provider=None, openai_endpoint_ur
     if get_db_setting("llm.supports_max_tokens", True):
         # Use 80% of context window to leave room for prompts
         max_tokens = min(
-            get_db_setting("llm.max_tokens", 30000), int(context_window_size * 0.8)
+            int(get_db_setting("llm.max_tokens", 30000)), int(context_window_size * 0.8)
         )
         common_params["max_tokens"] = max_tokens
